move Report class and rename to ExcelReport
